+++ cgkit/data/card_data.py
@@ -1,4 +1,4 @@
-from data import Data
+from .data import Data
 
 class CardData(Data):
     pass
